pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "djangorestsearch"
description = "ElasticSearch integration for Django"
readme = "README.rst"
requires-python = ">=3.9"
license = "BSD-2-Clause"
authors = [
    { name = "Jeremy Lainé", email = "jeremy.laine@m4x.org" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "aws-requests-auth>=0.3.0",
    "elasticsearch>=7.0.0,<8.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "celery",
    "coverage[toml]>=7.2.2",
    "django",
    "djangorestframework",
    "inflection",
    "ruff",
    "uritemplate",
]

[project.urls]
Homepage = "https://github.com/wemap/django-rest-search"
Documentation = "https://django-rest-search.readthedocs.io/"

[tool.coverage.run]
source = ["rest_search"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "W",  # pycodestyle
    "I",  # isort
    "T20",  # flake8-print
]

[tool.setuptools.dynamic]
version = {attr = "rest_search.__version__"}
